Move cmd_test.System type into internal/relayertest package
This will enable using the System type in the integration tests, which
cannot import from a different _test package.
